+++ R/border.R
@@ -20,6 +20,6 @@
 #' p + border()
 #'@export
 border <- function(color = "black", size = 0.8, linetype = NULL){
-   theme(panel.background = element_rect(color = color, size = size, linetype = linetype),
+   theme(panel.background = element_rect(color = color, linewidth = size, linetype = linetype),
             axis.line = element_blank())
 }