+++ R/geom_bracket.R
@@ -49,7 +49,8 @@ StatBracket <- ggplot2::ggproto("StatBracket", ggplot2::Stat,
 #' @param xmax numeric vector with the positions of the right sides of the
 #'   brackets
 #' @param y.position numeric vector with the y positions of the brackets
-#' @param size change the width of the lines of the bracket
+#' @param size change the width of the lines of the bracket. Deprecated, use \code{linewidth} instead.
+#' @param linewidth change the width of the lines of the bracket
 #' @param label.size change the size of the label text
 #' @param family change the font used for the text
 #' @param vjust move the text up or down relative to the bracket
@@ -140,7 +141,7 @@ stat_bracket <- function(mapping = NULL, data = NULL,
                          label = NULL, type = c("text", "expression"), y.position=NULL, xmin = NULL, xmax = NULL,
                          step.increase = 0, step.group.by = NULL,  tip.length = 0.03,
                          bracket.nudge.y = 0, bracket.shorten = 0,
-                         size = 0.3, label.size = 3.88, family="", vjust = 0,
+                         size = 0.3, linewidth = size, label.size = 3.88, family="", vjust = 0,
                          ...) {
   if(! is.null(data) & ! is.null(mapping)){
     if(! "x" %in% names(data)) mapping$x <- 1
@@ -155,7 +156,7 @@ stat_bracket <- function(mapping = NULL, data = NULL,
       y.position=y.position,xmin=xmin, xmax=xmax,
       step.increase=step.increase, bracket.nudge.y = bracket.nudge.y,
       bracket.shorten = bracket.shorten, step.group.by = step.group.by,
-      tip.length=tip.length, size=size, label.size=label.size,
+      tip.length=tip.length, size=linewidth, label.size=label.size,
       family=family, vjust=vjust, na.rm = na.rm, ...)
   )
 }
@@ -165,7 +166,7 @@ GeomBracket <- ggplot2::ggproto("GeomBracket", ggplot2::Geom,
                                 required_aes = c("x", "xend", "y", "yend", "annotation"),
                                 default_aes = ggplot2::aes(
                                   shape = 19, colour = "black", label.size = 3.88, angle = NULL, hjust = 0.5,
-                                  vjust = 0, alpha = NA, family = "", fontface = 1, lineheight = 1.2, linetype=1, size = 0.3,
+                                  vjust = 0, alpha = NA, family = "", fontface = 1, lineheight = 1.2, linetype=1, size = 0.3, linewidth = NULL,
                                   xmin = NULL, xmax = NULL, label = NULL, y.position = NULL, step.increase = 0,
                                   bracket.nudge.y = 0, bracket.shorten = 0 # Added to avoid aesthetics warning
                                   ),
@@ -212,7 +213,7 @@ GeomBracket <- ggplot2::ggproto("GeomBracket", ggplot2::Geom,
                                       gp = grid::gpar(
                                         col = scales::alpha(coords$colour, coords$alpha),
                                         lty = coords$linetype,
-                                        lwd = coords$size * ggplot2::.pt
+                                        lwd = (coords$linewidth %||% coords$size) * ggplot2::.pt
                                       )
                                     )
                                   )
@@ -227,7 +228,7 @@ geom_bracket <- function(mapping = NULL, data = NULL, stat = "bracket",
                          label = NULL, type = c("text", "expression"), y.position = NULL, xmin = NULL, xmax = NULL,
                          step.increase = 0, step.group.by = NULL, tip.length = 0.03,
                          bracket.nudge.y = 0, bracket.shorten = 0,
-                         size = 0.3, label.size = 3.88, family="", vjust = 0,
+                         size = 0.3, linewidth = size, label.size = 3.88, family="", vjust = 0,
                          coord.flip = FALSE,
                          ...) {
   type <- match.arg(type)
@@ -244,7 +245,7 @@ geom_bracket <- function(mapping = NULL, data = NULL, stat = "bracket",
     params = list(
       type = type,
       tip.length = tip.length,
-      size = size, label.size = label.size,
+      size = linewidth, label.size = label.size,
       family = family, na.rm = na.rm, coord.flip = coord.flip,
       ...
     )