+++ R/stat_regline_equation.R
@@ -115,7 +115,7 @@ stat_regline_equation <- function(
 
 StatReglineEquation<- ggproto("StatReglineEquation", Stat,
                   required_aes = c("x", "y"),
-                  default_aes = aes(label = ..eq.label.., hjust = ..hjust.., vjust = ..vjust..),
+                  default_aes = aes(label = after_stat(eq.label), hjust = after_stat(hjust), vjust = after_stat(vjust)),
 
                   compute_group = function(data, scales, formula, label.x.npc, label.y.npc,
                                            label.x, label.y, output.type, decreasing)